Simplify ifunc handling for relocatable executables
Turns out that we don't need relocations for our relocations! What I was
doing before worked, but it wasn't pretty. Turns out glibc has different
code paths for resolving ifuncs and we were using the non-PIE code path
in PIE executables and making it work by applying relocations to
the ifunc relocations. However if we just don't define the iplt
start/stop symbols, then the proper code path is used, which doesn't
need (and must not have) relocations applied to the ifunc relocations.