refactoring: Change the way relocations in .eh_frame are handled
Previously we computed a relocation action and stored it. Now we just
store a slice of relocations that need to be applied. This is possible
now that we're using the lower level object APIs. By doing this, we now
always apply a relocation action straight after computing it, which
opens the possibility of getting rid of relocation actions.