Apply addend after string-merge resolution for named symbols
Previously, we always applied the addend then determined which string in
the string-merge section we were referencing. However that doesn't match
the behaviour of GNU ld. It seems that for named symbols we need to
determine which string we're referencing and only then apply the addend.

For unnamed symbols (section references), we continue to apply the
addend before determining which string we're referencing.

Issue #24