linker-diff: Add support for various 64-bit relocation types
This change deviates somewhat from how linker-diff has matched and
resolved instructions in the past. It's possible that going forward,
more of how it works will move over to this way of working.

The new approach works better with relocations that have offsets and
with PC-relative relocations but where the instruction isn't
PC-relative.

Issue #27