Return RelocationKindInfo when relaxing rather than a raw relocation type
This will allow relaxing to internal-only relocation kinds. I thought
I'd need this for the next commit - turns out that I didn't, but I'm
going to keep this anyway, since it gives us the flexibility should we
need it in future and doesn't really add any extra compexity. It also
means less use of raw u32 relocation types, which is nice.