Make basic lazy PLT stuff work
.got is now separated into .got and .got.plt.

.plt is now separated into .plt and .plt.got.

The naming of these sections is super-confusing, but matches what GNU ld does.

There's a bunch of things where we don't do the same as GNU ld. In particular,
we still generate .got.plt and .plt.got in cases when GNU ld doesn't. But
fixing those is left for later, since this change is big enough as-is.

ifuncs aren't yet done correctly, since it turns out it's hard to do that
without breaking ifuncs for non-pie static binaries.

Issue #70