Rewrite asm_diff.rs
Details about how the new implementation works are in comments in the code.
There's still lots more to be done, but the new approach should be a much more
solid foundation on which to build.

The new implementation relies on having a .layout file.

The new implmentation doesn't use a disassembler for diffing, only for
reporting problems. This should make porting to aarch64 much more practical,
since we no longer need to extract information from the disassembled
instructions.