Make relaxations adjust `place` when computing relocation value
This should be a no-op, but makes the code less confusing. Only one relaxation
both adjusted the offset and was used in conjuction with a place-relative
relocation. That relaxation, TlsGdToInitialExec now doesn't need to set a weird
addend of -12, which was only needed previously because the old offset was used
as the place.