Load SEC when __start_SEC / __stop_SEC is referenced
We previously only kept such sections if they had the retain bit set.

This should match the behaviour of GNU ld and also of lld when -z
nostart-stop-gc is set. We now treat `-z nostart-stop-gc` as a no-op rather
than warning.

The custom sections test is updated to not use the retain bit, since it isn't
needed anymore. This also allowed that test to enabled for cross compilation,
which was disabled due to the use of the retain bit.

Fixes: #459