Don't attempt to export dynamic symbols for discarded archive entries
At the place where we were requesting export of the symbol, we didn't have easy
access to whether the object defining the symbol was discarded, so the code is
changed to send an ExportDynamic request to the defining object. It can then
ignore the request if it wasn't loaded.

Fixes #498