Remove StorageModel
This was intended to allow us to store the symbol name hashmap on disk for use
by incremental linking, however I've decided, at least initially to keep the
incremental linking state in memory, which means we don't need this. While I
might decide to bring it back in the future, right now it's an unnecessary
abstraction that's kind of in the way.