Fixes to --time output
Start time from when the scope is entered rather than when the scope is
created. This mostly only affects the shutdown time, for which we currently
create the scope at the start of linking, but only later enter it.

We also move unmapping of the output file to the write phase rather than the
shutdown phase, since the output file isn't really usable until we've unmapped
it, since we'll still be holding an exclusive lock on the file.