Reduce layers of borrowing
From the perspective of the command-line interface, this shouldn't change any
observable behaviour.

From the perspective of the libwild API, it's substantially altered.

For input data and groups, we now keep the prelude and epilogue more separate
than we did before. This helps because prelude in particular has different
lifetime constraints than regular object files.

We make use of a typed arena to hold our input data. This allows us to read it,
then borrow from it.

Another strategy that's employed is passing ownership to the next stage rather
than having the next stage borrow the previous stage.

The end effect is that we now have a `Linker` struct and running the linker
borrows from `Linker` and from `Args` and returns a `LinkerOuput`. This means
that cleanup can be done outside of the main link function, which means that we
no longer need to pass a `done_callback` into linking.