Replace CAN_BYPASS_GOT with NON_INTERPOSABLE
While there is a lot of overlap between these two concepts, they're not exactly
the same. There are thread-locals where we cannot bypass the GOT even though
they're local variables and thus non-interposable.

In a later change, we also want to generate certain runtime relocations for
non-interposable symbols in ways that don't involve the GOT.