Fix linking against protected symbols in shared objects
There were two problems. Firstly, we were copying the visibility of the symbol
from the shared object to the executable, which meant that the runtime loader
would ignore relocations for that symbol.

Secondly, we were marking the symbol as NON_INTERPOSABLE, which doesn't really
make sense when the symbol is dynamic, so coming from elsewhere.

Issue #668