Set section flags by merging flags of input sections
This is necessary for linker scripts, since otherwise we don't really know what
flags to put on the sections specified by the linker script.

It meant that we had to decide the the ordering later, since we need to know
section flags in order to know where to put sections.

Issue #44