Remove the need for output sections to know their loadable segment
This was a bit ugly before because the loadable_segment_id was initially None,
then filled in later.

The only thing it was used for was to know when we were starting a new loadable
segment, which is much more cleanly found by just iterating through the order
events.

In order to directly access the order events, we skip calling
OutputSectionPartMap::output_order_map, so that was refactored slightly.