copy-relocations test: Fix when running with qemu
Since we don't yet run with lld when running with qemu, we can't rely on
matching LLD's behaviour, so we need to ignore a case where GNU ld doesn't
export a weak symbol alias and we and LLD do.

Also, rename various things in the test to make them clearer.

Fixes #705