linker-diff: Don't resolve referent when relocation is eliminated
This was mostly an issue for TlsLdToLocalExec.

A TLSLD relocation is really a reference to the module containing the symbol,
not to the symbol itself, so trying to extract a symbol from it will lead to
false diffs. When a relaxation converts the relocation type to none, we now
don't try to resolve a symbol.

Issue #745