add a docker reproducer for nix
This is based on the solutions provided in #832.

The Nix shell includes tools needed for fixing tests as well as rustup
and cargo-chef for Docker, as `nix-shell` does not seem to support
overlaying both the `shell.nix` and packages specified by CLI.

The use of rustup in the Nix shell/Dockerfile is not the way this would
typically be done in Nix; most use a project like
https://github.com/oxalica/rust-overlay to handle downloading and using
Rust. However, rust-overlay seems to do some sort of patching and
wrapping that leads to even more test failures, so for now we stick to
rustup.