Change behaviour of --no-allow-shlib-undefined to match lld
It turns out that LLD is more relaxed about undefined symbols in shared objects
compared to GNU ld. We now hopefully do the same relaxed behaviour as LLD. That
is, we only report an error if the shared object has all of its direct
dependencies in the link.

Also, change the default back to --no-allow-shlib-undefined.

Fixes #897