Move per-symbol-flags out of SymbolDb
This allows us to transition between atomic and non-atomic access in a less
error-prone way. Specifically, we no longer have to worry about calling methods
on SymbolDb that need access to the value flags, but find them absent resulting
in a panic. Instead, such methods now take the value flags as an argument, so
can't build unless we can pass them.