refactoring: Change resolution to use a work queue (#1290)
This has negligible affect on any performance metrics.

Before we had a rayon scope into which we spawned tasks. Inside those
tasks we did further rayon parallelism with par_iter.

Now, we just have a single level of parallelism which pulls from a
channel. Large objects are split into separate work items in that
channel.

By not using nested parallelism, it should be easier to experiment with
different parallelism libraries.