Nix: Update for inclusion in Nixpkgs (#1318)
[Wild is now in
Nixpkgs](https://search.nixos.org/packages?channel=unstable&show=wild&query=wild),
which is great. So in this PR I move some stuff around because of that.

1. Bump the lock file
2. Have the overlay set `wild-unwrapped` as Wild needs to be wrapped to
find libraries correctly. The wrapper is in Nixpkgs.
3. Run the tests that are in Nixpkgs (these are the same tests as were
in the Wild repo before)
4. Remove the stdenv adapter from the Wild repo
5. Remove the two seperate overlay files and centralize to just one
6. Move the GCC wrappers to their own file

Then there is a commit that does several things all at once since
there's not a great way to split most of them up:
1. Ensure all accesses of Wild are done through the overlay
2. Move the check inputs into the derivation
3. Remove some of the hacks for running unit tests by using `stdenvNoCC`
when running unit tests