refactor: Move dynamic symbol definitions onto layout
The epilogue still allocates space for these and writes them, but it's
convenient to just store them on the layout. This eliminates one bit of
code where we have to get access to the epilogue. Perhaps more
importantly, it changes merge_dynamic_symbol_definitions to only take
shared state, which means it could in future be run in parallel with
something else.